% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addContour.R
\name{addContour}
\alias{addContour}
\title{Add Contour to Section Plot}
\usage{
addContour(
  section,
  field = NULL,
  levels = NULL,
  col = "black",
  lty = 1,
  lwd = 1,
  labels = NULL,
  cex.lab = 1
)
}
\arguments{
\item{section}{a section object from "build.section()"}

\item{field}{the field name to use for contouring}

\item{levels}{the level(s) to draw contours at}

\item{col}{the color of the contour line(s)}

\item{lty}{the lty for the line(s)}

\item{labels}{a boolean for whether the label(s) should be drawn}

\item{cex.lab}{the cex (i.e. size) of the labels}
}
\description{
Add Contour to Section Plot
}
\author{
Thomas Bryce Kelly
}
