% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMask.R
\name{addMask}
\alias{addMask}
\title{Add Section mask}
\usage{
addMask(section, bathy = bathy.global, binning = 1, bathy.col = "darkgrey")
}
\arguments{
\item{section}{a section object from build.section()}

\item{bathy}{a bathymetric object such as "bathy.global" (the default)}

\item{binning}{a smoothing paramter; must be an odd integer}

\item{bathy.col}{the color of the bathymetry}
}
\description{
Add Section mask
}
\author{
Thomas Bryce Kelly
}
