% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGrid.old.R
\name{buildGrid.old}
\alias{buildGrid.old}
\title{Build Grid}
\usage{
buildGrid.old(
  x,
  y,
  z,
  nx = 50,
  ny = 50,
  x.scale = NULL,
  y.scale = NULL,
  gridder = gridWeighted,
  grid = NULL,
  xlim = NULL,
  ylim = NULL,
  x.factor = NULL,
  y.factor = NULL,
  neighborhood = 20
)
}
\description{
Build Grid
}
\author{
Thomas Bryce Kelly
}
\keyword{Gridding}
