% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGrid.R
\name{buildGrid}
\alias{buildGrid}
\title{Build Grid}
\usage{
buildGrid(
  xlim,
  ylim,
  nx = NULL,
  ny = NULL,
  x.scale = NULL,
  y.scale = NULL,
  x.factor = NULL,
  y.factor = NULL
)
}
\description{
Build Grid
}
\author{
Thomas Bryce Kelly
}
\keyword{Gridding}
