% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridNN.R
\name{gridNN}
\alias{gridNN}
\title{Grid via Nearest Neighbor}
\usage{
gridNN(x, y, z, tree, gx, gy, dx, dy, neighborhood = NULL, func = NULL)
}
\arguments{
\item{x}{Observations, x values}

\item{y}{Observations, y values}

\item{z}{Observations, z values}

\item{gx}{Grid x values to interpolate onto}

\item{gy}{Grid y values to interpolate onto}

\item{p}{Unused}

\item{xscale}{The spacing of the x grid}

\item{yscale}{The spacing of the y grid}

\item{uncertainty}{Unused}
}
\description{
Bins data into reguarly spaced grid
}
\author{
Thomas Bryce Kelly
}
