% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridWeighted.R
\name{gridWeighted}
\alias{gridWeighted}
\title{Grid via Inverse Distance Weighting Interpolation}
\usage{
gridWeighted(
  gx,
  gy,
  x,
  y,
  z,
  p = 2,
  xscale = 1,
  yscale = 1,
  uncertainty = 0.1,
  neighborhood = NULL,
  x.factor = 1,
  y.factor = 1,
  verbose = F
)
}
\arguments{
\item{gx}{Grid x values to interpolate onto}

\item{gy}{Grid y values to interpolate onto}

\item{x}{Observations, x values}

\item{y}{Observations, y values}

\item{z}{Observations, z values}

\item{p}{Exponent on the distance function}
}
\description{
Grid via Inverse Distance Weighting Interpolation
}
\author{
Thomas Bryce Kelly
}
