% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridWeighted.R
\name{gridWeighted}
\alias{gridWeighted}
\title{Grid via Inverse Distance Weighting Interpolation}
\usage{
gridWeighted(
  x,
  y,
  z,
  tree,
  gx,
  gy,
  dx,
  dy = NULL,
  neighborhood = 25,
  func = function(x) {
     1/x^2
 }
)
}
\arguments{
\item{x}{Observations, x values}

\item{y}{Observations, y values}

\item{z}{Observations, z values}

\item{gx}{Grid x values to interpolate onto}

\item{gy}{Grid y values to interpolate onto}

\item{p}{Exponent on the distance function}
}
\description{
Grid via Inverse Distance Weighting Interpolation
}
\author{
Thomas Bryce Kelly
}
