% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Plot Grid}
\usage{
plotGrid(
  section,
  xlim = NULL,
  ylim = NULL,
  xlab = "x",
  ylab = "y",
  zlim = NULL,
  pal = "greyscale",
  rev = F,
  N = 16,
  ...
)
}
\description{
Plot a section object.
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
\keyword{Section}
